Extended las2las tool with -save_vlr and load_vlr options for exporting and importing a single VLR.
The tool now supports reading an individual VLR from the file header and save it into a vlr file using the save_vlr function.
You can specify the VLR by its index or user ID and record ID, and optionally, the vlr filename for saving.

The tool now also supports loading an individual VLR from the vlr file and adding it to the file header using the load_vlr function.
You can specify the VLR by its index or user ID and record ID, and optionally provide a filename for the vlr file.