Adjustments to the PROJ library implementation in LAStools:
The global encoding bit for WKT is now always set to ‘true’ when a PROJ transformation is performed. All VLR entries from the source file are not transferred to the target, only the WKT is written to the target.

A hierarchy has been implemented to determine the source CRS for the PROJ transformation:
 The source CRS is passed as an argument.
 If not available, a WKT is searched for in the header of the source file.
 If no WKT is available, it is generated from the GeoTIFF data.
 If this is not possible, the EPSG code from the GeoTIFF is used, which can lead to inaccuracies as GeoTIFF arguments may be ignored.

The WKT generated by the PROJ library is now returned without breaks and indentations.

Additional LAS_VERBOSE and LAS_VERY_VERBOSE outputs have been added.