Fix int32_t overflow causing lascopcindex to crash
Seeing a malloc for 44,481,125points * 50bytes = 2,224,056,250 bytes in load(), which won't fit into the int32_t you're using. This casts to a crazy large size_t and then causes malloc to return a nullptr.

Fix is to use size_t, which is what malloc expects.

I do not know why its trying to read 44M points at a time and why this only ever crashed on linux but worked on windows.

This might be rapidlasso support issue 653.